requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
email_validator==2.2.0
Faker==33.0.0
greenlet==3.1.1
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
